+++ src/hud/Hud.lua
@@ -42,7 +42,7 @@ function Hud.New()
 
                 local html = tpl({
                     throttle = player.isFrozen() and
-                        string.format("Throttle: %0.1f (manual control active)", input.Throttle() * 100) or "",
+                        string.format("Throttle: %0.0f%% (manual control active)", input.Throttle() * 100) or "",
                     fuelByType = fuelByType,
                     isECU = isECU,
                 })