+++ src/flight/FlightCore.lua
@@ -26,7 +26,7 @@ require("flight/state/Require")
 ---@field Turn fun(degrees:number, axis:Vec3):Vec3
 ---@field AlignTo fun(point:Vec3)
 ---@field StopEvents fun()
----@field CreateWPFromPoint fun(p:Point, lastInRoute:boolean):Waypoint
+---@field CreateWPFromPoint fun(p:Point, lastInRoute:boolean, pathAlignmentDistanceLimitFromSurface:number):Waypoint
 ---@field GoIdle fun()
 ---@field GotoTarget fun(target:Vec3, lockdir:Vec3, margin:number, maxSpeed:number, finalSpeed:number, ignoreLastInRoute:boolean, forceVerticalUp:boolean)
 ---@field WaitForGate fun():boolean
@@ -41,8 +41,9 @@ local defaultMargin = constants.flight.defaultMargin
 ---Creates a waypoint from a point
 ---@param point Point
 ---@param lastInRoute boolean
+---@param pathAlignmentDistanceLimitFromSurface number
 ---@return Waypoint
-function FlightCore.CreateWPFromPoint(point, lastInRoute)
+function FlightCore.CreateWPFromPoint(point, lastInRoute, pathAlignmentDistanceLimitFromSurface)
     local opt = point.Options()
     local lockDir = Vec3.New(opt.Get(PointOptions.LOCK_DIRECTION, Vec3.zero))
     local margin = opt.Get(PointOptions.MARGIN, defaultMargin)
@@ -56,7 +57,7 @@ function FlightCore.CreateWPFromPoint(point, lastInRoute)
 
     local coordinate = universe.ParsePosition(point.Pos()).Coordinates()
 
-    local wp = Waypoint.New(coordinate, finalSpeed, maxSpeed, margin)
+    local wp = Waypoint.New(coordinate, finalSpeed, maxSpeed, margin, pathAlignmentDistanceLimitFromSurface)
     wp.SetLastInRoute(lastInRoute)
 
     if lockDir ~= Vec3.zero then
@@ -86,7 +87,7 @@ function FlightCore.New(routeController, flightFSM)
     local routePublishTimer = Stopwatch.New()
 
     local function createDefaultWP()
-        return Waypoint.New(vehicle.position.Current(), 0, 0, defaultMargin)
+        return Waypoint.New(vehicle.position.Current(), 0, 0, defaultMargin, 0)
     end
 
     -- Setup start waypoints to prevent nil values
@@ -112,7 +113,8 @@ function FlightCore.New(routeController, flightFSM)
         end
 
         previousWaypoint = currentWaypoint
-        currentWaypoint = FlightCore.CreateWPFromPoint(nextPoint, route.LastPointReached())
+        currentWaypoint = FlightCore.CreateWPFromPoint(nextPoint, route.LastPointReached(),
+            settings.Number("pathAlignmentDistanceLimitFromSurface"))
 
         -- When the next waypoint is nearly above or below us, lock yaw
         local dir = (currentWaypoint.Destination() - previousWaypoint.Destination()):NormalizeInPlace()