+++ src/info/Fuel.lua
@@ -5,8 +5,9 @@ local Task             = require("system/Task")
 local Vec2             = require("native/Vec2")
 local log              = require("debug/Log").Instance()
 local pub              = require("util/PubSub").Instance()
+local _                = require("util/Table")
 
----@alias FuelTankInfo {name:string, factorBar:Vec2, percent:number, visible:boolean}
+---@alias FuelTankInfo {name:string, factorBar:Vec2, percent:number, visible:boolean, type:string}
 
 ---@class Fuel
 ---@field Instance fun():Fuel
@@ -22,7 +23,13 @@ function Fuel.New(settings)
         return instance
     end
 
-    local talents = ContainerTalents.New(0, 0, 0, 0, 0, 0)
+    local talents = ContainerTalents.New(
+        settings.Number("containerProficiency", 0),
+        settings.Number("fuelTankOptimization", 0),
+        settings.Number("containerOptimization", 0),
+        settings.Number("atmoFuelTankHandling", 0),
+        settings.Number("spaceFuelTankHandling", 0),
+        settings.Number("rocketFuelTankHandling", 0))
 
     local s = {}
 
@@ -30,7 +37,7 @@ function Fuel.New(settings)
         local sw = Stopwatch.New()
         sw.Start()
 
-        local tanks = {
+        local fuelTanks = {
             atmo = Container.GetAllCo(ContainerType.Atmospheric),
             space = Container.GetAllCo(ContainerType.Space),
             rocket = Container.GetAllCo(ContainerType.Rocket)
@@ -40,34 +47,50 @@ function Fuel.New(settings)
             if sw.IsRunning() and sw.Elapsed() < 2 then
                 coroutine.yield()
             else
-                for fuelType, containers in pairs(tanks) do
-                    local fillFactors = {} ---@type FuelTankInfo[]
-                    for _, tank in ipairs(containers) do
+                local byType = {} ---@type table<string,FuelTankInfo[]>
+
+                for fuelType, tanks in pairs(fuelTanks) do
+                    for _, tank in ipairs(tanks) do
                         local factor = tank.FuelFillFactor(talents)
-                        table.insert(fillFactors,
-                            {
-                                name = tank.Name(),
-                                factorBar = Vec2.New(1, factor),
-                                percent = factor * 100,
-                                visible = true
-                            })
-                        coroutine.yield()
+                        local curr = {
+                            name = tank.Name(),
+                            factorBar = Vec2.New(1, factor),
+                            percent = factor * 100,
+                            visible = true,
+                            type = fuelType
+                        }
+
+                        local bt = byType[fuelType] or {}
+                        bt[#bt + 1] = curr
+                        byType[fuelType] = bt
                     end
 
-                    -- Sort tanks in acending fuel levels
-                    table.sort(fillFactors,
-                        function(a, b) return a.percent < b.percent end)
-
-                    local fuelData = {}
-
-                    for i, tankInfo in ipairs(fillFactors) do
-                        fuelData[#fuelData + 1] = { path = string.format("fuel/%s/%d", fuelType, i), tank = tankInfo }
+                    -- Sort tanks for HUD in alphabetical order
+                    if byType[fuelType] then
+                        table.sort(byType[fuelType], function(a, b)
+                            return a.name < b.name
+                        end)
                     end
-
-                    pub.Publish("FuelData", fuelData)
-
-                    coroutine.yield()
                 end
+
+                pub.Publish("FuelByType", DeepCopy(byType))
+
+                -- Now populate for the screen
+                local fuelForScreen = {} ---@type {path:string, tank:FuelTankInfo}[]
+
+                -- Sort tanks for screen based on acending fuel levels
+                for fuelType, tanks in pairs(byType) do
+                    table.sort(tanks, function(a, b) return a.percent < b.percent end)
+
+                    for i, curr in ipairs(tanks) do
+                        fuelForScreen[#fuelForScreen + 1] = {
+                            path = string.format("fuel/%s/%d", fuelType, i),
+                            tank = curr
+                        }
+                    end
+                end
+
+                pub.Publish("FuelData", fuelForScreen)
 
                 sw.Restart()
             end